Added project files for VC8.0.
git-svn-id: http://www.nflc.org/svnrepos/newauthtool/branches/php_xydiff@779 c3f307ee-532b-8d43-9a74-367e9d9c948c

+++ xydiff_error_handler.cpp
@@ -87,8 +87,8 @@ void xydiffPHPParseHandler::fatalError(const SAXParseException& e) {
 }
 void xydiffPHPParseHandler::warning(const SAXParseException& e) {
 	std::cerr << "\n(GF) Warning at (file " << XMLString::transcode(e.getSystemId())
-	<< ", line " << e.getLineNumber()
-	<< ", char " << e.getColumnNumber()
+	<< ", line " << (long) e.getLineNumber()
+	<< ", char " << (long) e.getColumnNumber()
 	<< "): " << XMLString::transcode(e.getMessage()) << std::endl;
 }
